requirements.txt
supervisely==6.23.1
google-cloud-storage==1.35.0